+++ packages/visualizations/src/index.ts
@@ -4,6 +4,7 @@ export { default as KpiCard } from './components/KpiCard';
 export { ChoroplethGeoJson, ChoroplethVectorTiles } from './components/Map/WebGl';
 export { default as ChoroplethSvg } from './components/Map/Svg';
 export { default as PoiMap } from './components/MapPoi';
+export { default as Table } from './components/Table';
 
 export * from './types';
 export * from './components/types';
@@ -13,6 +14,7 @@ export * from './components/Map/types';
 export * from './components/MarkdownText/types';
 export * from './components/MapPoi/types';
 export * from './components/Legend/types';
+export * from './components/Table/types';
 
 /* We export Svelte types from the main package to avoid conflicting versions
 It still is purely a dev thing and we'll bundle they types directly */